- moved the function which returns the required data to service layer and the corresponding queries to model. - removed test cases from integration testing and moved it to unit testing. - moved updateTaskStatusToDone to tasks model as it was causing circular dependency. - created a separate fixture file for mock data required to test the feature. - reverted the test case changes done for other test cases due to mock data changes.
