+++ test/fixtures/user/user.js
@@ -442,68 +442,5 @@ module.exports = () => {
         url: "https://res.cloudinary.com/realdevsquad/image/upload/v1667685133/profile/mtS4DhUvNYsKqI7oCWVB/aenklfhtjldc5ytei3ar.jpg",
       },
     },
-    {
-      id: "user1_id",
-      discordId: "123456789",
-      github_id: "github_user1",
-      username: "archived_user1",
-      first_name: "Archived",
-      last_name: "User One",
-      linkedin_id: "archived_user1",
-      github_display_name: "archived-user-1",
-      phone: "1234567890",
-      email: "archived1@test.com",
-      roles: {
-        archived: true,
-        in_discord: false,
-      },
-      discordJoinedAt: "2024-01-01T00:00:00.000Z",
-      picture: {
-        publicId: "profile/user1",
-        url: "https://example.com/user1.jpg",
-      },
-    },
-    {
-      id: "user2_id",
-      discordId: "987654321",
-      github_id: "github_user2",
-      username: "archived_user2",
-      first_name: "Archived",
-      last_name: "User Two",
-      linkedin_id: "archived_user2",
-      github_display_name: "archived-user-2",
-      phone: "0987654321",
-      email: "archived2@test.com",
-      roles: {
-        archived: true,
-        in_discord: false,
-      },
-      discordJoinedAt: "2024-01-02T00:00:00.000Z",
-      picture: {
-        publicId: "profile/user2",
-        url: "https://example.com/user2.jpg",
-      },
-    },
-    {
-      id: "user3_id",
-      discordId: "555555555",
-      github_id: "github_user3",
-      username: "active_user",
-      first_name: "Active",
-      last_name: "User",
-      linkedin_id: "active_user",
-      github_display_name: "active-user",
-      phone: "5555555555",
-      email: "active@test.com",
-      roles: {
-        archived: false,
-        in_discord: true,
-      },
-      discordJoinedAt: "2024-01-03T00:00:00.000Z",
-      picture: {
-        publicId: "profile/user3",
-        url: "https://example.com/user3.jpg",
-      },
-    },
   ];
 };
