+++ test/integration/tasks.test.js
@@ -1633,62 +1633,4 @@ describe("Tasks", function () {
       });
     });
   });
-
-  describe("fetchOrphanedTasks", function () {
-    beforeEach(async function () {
-      // Clean the database
-      await cleanDb();
-
-      // Add test users to the database
-      const userPromises = userData.map((user) => userDBModel.add(user));
-      await Promise.all(userPromises);
-
-      // Add test tasks to the database
-      const taskPromises = tasksData.map((task) => tasksModel.add(task));
-      await Promise.all(taskPromises);
-    });
-
-    afterEach(async function () {
-      await cleanDb();
-    });
-
-    it("should fetch tasks assigned to archived and non-discord users", async function () {
-      const abandonedTasks = await tasks.fetchOrphanedTasks();
-
-      expect(abandonedTasks).to.be.an("array");
-      expect(abandonedTasks).to.have.lengthOf(2); // Two tasks abandoned by users
-    });
-
-    it("should not include completed or done tasks", async function () {
-      const abandonedTasks = await tasks.fetchOrphanedTasks();
-
-      abandonedTasks.forEach((task) => {
-        expect(task.status).to.not.be.oneOf(["DONE", "COMPLETED"]);
-      });
-    });
-
-    it("should not include tasks from active users", async function () {
-      const abandonedTasks = await tasks.fetchOrphanedTasks();
-
-      abandonedTasks.forEach((task) => {
-        expect(task.assignee).to.not.equal("active_user");
-      });
-    });
-
-    it("should handle case when no users are archived", async function () {
-      await cleanDb();
-
-      // Add only active users
-      const activeUser = userData[11]; // Using the active user from our test data
-      await userDBModel.add(activeUser);
-
-      // Add a task assigned to the active user
-      const activeTask = tasksData[11]; // Using the active user's task
-      await tasksModel.add(activeTask);
-
-      const abandonedTasks = await tasks.fetchOrphanedTasks();
-      expect(abandonedTasks).to.be.an("array");
-      expect(abandonedTasks).to.have.lengthOf(0);
-    });
-  });
 });
